requirements.txt
streamlit
python-docx
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
openai
python-dotenv
pandas
numpy
faiss-cpu
google-generativeai